requirements.txt
Flask
flask-bcrypt
flask-cors
sqlite
